examples/smart-forgetfulness-care/pyproject.toml
[project]
name = "smart-forgetfulness-care"
version = "0.1.0"
description = "Outing Guardian Agent: Remote Care for Forgetfulness"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.3.5",
    "langchain>=1.2.3",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.3.0",
    "langgraph>=1.0.6",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "langfuse>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
